UIの変更
残り時間を数字表記から時計に変更した
現在の目標を追加した

+++ Christmas-Gift/Project/GameProject/GameScene/GameData.cpp
@@ -11,6 +11,6 @@ bool GameData::StartFadeOut = false;
 bool GameData::GameStart = false;
 bool GameData::isPauseGame = false;
 
-int GameData::second = 0;					//タイマー
+float GameData::second = 0;					//タイマー
 
 float GameData::FacePosition = 340;			//顔UI配置
